test: split PDF bibliography continuation cases
